move some classes to test/utils/.
